+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Fix empty search string handling.
+- Prevent the query builder from handling empty query variables.
 
 ## [1.1.1] - 2020-02-06